detector/requirements.txt
opencv-python
numpy
ultralytics
paho-mqtt

